Combine G-code files into fewer units
Addressing #7707
