--- Marlin/src/gcode/feature/mixing/M164.cpp
@@ -1,45 +1,0 @@
-/**
- * Marlin 3D Printer Firmware
- * Copyright (C) 2016 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
- *
- * Based on Sprinter and grbl.
- * Copyright (C) 2011 Camiel Gubbels / Erik van der Zalm
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-#include "../../../inc/MarlinConfig.h"
-
-#if ENABLED(MIXING_EXTRUDER) && MIXING_VIRTUAL_TOOLS > 1
-
-#include "../../gcode.h"
-#include "../../../feature/mixing.h"
-
-/**
- * M164: Store the current mix factors as a virtual tool.
- *
- *   S[index]   The virtual tool to store
- *
- */
-void GcodeSuite::M164() {
-  const int tool_index = parser.intval('S');
-  if (tool_index < MIXING_VIRTUAL_TOOLS) {
-    normalize_mix();
-    for (uint8_t i = 0; i < MIXING_STEPPERS; i++)
-      mixing_virtual_tool_mix[tool_index][i] = mixing_factor[i];
-  }
-}
-
-#endif // MIXING_EXTRUDER && MIXING_VIRTUAL_TOOLS > 1
