--- Marlin/src/gcode/stats/M75-M77.cpp
@@ -1,39 +1,0 @@
-/**
- * Marlin 3D Printer Firmware
- * Copyright (C) 2016 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
- *
- * Based on Sprinter and grbl.
- * Copyright (C) 2011 Camiel Gubbels / Erik van der Zalm
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-#include "../gcode.h"
-#include "../../module/printcounter.h"
-
-/**
- * M75: Start print timer
- */
-void GcodeSuite::M75() { print_job_timer.start(); }
-
-/**
- * M76: Pause print timer
- */
-void GcodeSuite::M76() { print_job_timer.pause(); }
-
-/**
- * M77: Stop print timer
- */
-void GcodeSuite::M77() { print_job_timer.stop(); }
