--- Marlin/src/gcode/stats/M78.cpp
@@ -1,40 +1,0 @@
-/**
- * Marlin 3D Printer Firmware
- * Copyright (C) 2016 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
- *
- * Based on Sprinter and grbl.
- * Copyright (C) 2011 Camiel Gubbels / Erik van der Zalm
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-#include "../../inc/MarlinConfig.h"
-
-#if ENABLED(PRINTCOUNTER)
-
-#include "../gcode.h"
-#include "../../module/printcounter.h"
-
-/**
- * M78: Show print statistics
- */
-void GcodeSuite::M78() {
-  if (parser.intval('S') == 78)   // "M78 S78" will reset the statistics
-    print_job_timer.initStats();
-  else
-    print_job_timer.showStats();
-}
-
-#endif // PRINTCOUNTER
